sm8350-common: Ship prebuilt android.frameworks.sensorservice from op5daal1
From: qssi_64-user 15 AP3A.240617.008 1730816783861 release-keys
ref: https://dumps.tadiphone.dev/dumps/oneplus/op5daal1

I hwservicemanager: Since android.frameworks.sensorservice@1.0::ISensorManager/default is not registered, trying to start it as a lazy HAL (if it's not configured to be a lazy HAL, it may be stuck starting or still starting).
I HidlServiceManagement: getService: Trying again for android.frameworks.sensorservice@1.0::ISensorManager/default...
W libc    : Unable to set property "ctl.interface_start" to "android.frameworks.sensorservice@1.0::ISensorManager/default": PROP_ERROR_HANDLE_CONTROL_MESSAGE (0x20)
I hwservicemanager: Tried to start android.frameworks.sensorservice@1.0::ISensorManager/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Change-Id: 5aab451fe6e3698bbdb6fbc435c0564a5675c51d
Signed-off-by: TogoFire <togofire@mailfence.com>